pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.1",
    "django-bootstrap5>=25.1",
    "django-filter>=25.1",
    "gunicorn>=23.0.0",
    "python-dotenv>=1.1.0",
    "ruff>=0.11.10",
]
